requirements.txt
# Core dependencies
einops==0.7.0
numpy==1.26.1
scipy==1.11.3
timm==0.6.13
huggingface-hub
soundfile

# Noise Reduction
noisereduce>=2.0.0
webrtcvad-wheels>=2.0.10

# ONNX Runtime (Default CPU)
# For GPU support: pip install onnxruntime-gpu
onnxruntime>=1.16.0

# Optional: PyTorch (only if using original model loading)
# torch
